+++ model/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/model/value/primitive/FloatValue.java
@@ -39,7 +39,7 @@ public class FloatValue extends TypedValue<Float> {
             return;
         }
         try {
-            this.setValue((Float) Float.parseFloat(value));
+            this.setValue(Float.parseFloat(value));
         }
         catch (NumberFormatException ex) {
             throw new ValueFormatException(ex);
