+++ model/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/model/value/primitive/ShortValue.java
@@ -39,7 +39,7 @@ public class ShortValue extends TypedValue<Short> {
             return;
         }
         try {
-            this.setValue((Short) Short.parseShort(value));
+            this.setValue(Short.parseShort(value));
         }
         catch (NumberFormatException ex) {
             throw new ValueFormatException(ex);
